Now with menu and a scene file
Hi guys, I made a basic start and end screen and hooked it up to a lvl
manager so you can actually go to the game from the start menu or to the
menu or game from the end scene. an engineer will need to hook up the
dying to trigger end scene part. Also I made a folder to put the scenes
into since there are now multiple.